==> Exploratory Data Analysis - II/Notebook_1 Activity_0.ipynb <==
## Dplyr Vignette 

library(dplyr)
library(nycflights13)

my_db <- src_sqlite("my_db_f1.sqlite3", create = T)

flights_sqlite <- copy_to(my_db, flights, temporary = FALSE, indexes = list(
  c("year", "month", "day"), "carrier", "tailnum"))

flights_sqlite <- tbl(nycflights13_sqlite(), "flights")
flights_sqlite

tbl(my_db, sql("SELECT * FROM flights"))

select(flights_sqlite, year:day, dep_delay, arr_delay)

filter(flights_sqlite, dep_delay > 240)

mutate(flights_sqlite, speed = air_time / distance)

summarise(flights_sqlite, delay = mean(dep_time))

c1 <- filter(flights_sqlite, year == 2013, month == 1, day == 1)
c2 <- select(c1, year, month, day, carrier, dep_delay, air_time, distance)
c3 <- mutate(c2, speed = distance / air_time * 60)
c4 <- arrange(c3, year, month, day, carrier)

c4

explain(c4)

translate_sql(x)

translate_sql("x")

translate_sql(x == 1 && (y < 2 || z > 3))

translate_sql(x ^ 2 < 10)

translate_sql(x %% 2 == 10)

translate_sql(1)

translate_sql(1L)

translate_sql(glob(x, y))
#> <SQL> GLOB("x", "y")
translate_sql(x %like% "ab*")
#> <SQL> "x" LIKE 'ab*'

by_tailnum <- group_by(flights_sqlite, tailnum)
delay <- summarise(by_tailnum,
  count = n(),
  dist = mean(distance),
  delay = mean(arr_delay)
)
delay <- filter(delay, count > 20, dist < 2000)
delay_local <- collect(delay)
