==> Exploratory Data Analysis - II/Notebook_2 Activity_2.ipynb <==
## Loading Libraries

library(dplyr)
my_db <- src_sqlite("database.sqlite", create = T)

## Converting Tables into Dataframes

player_attribute <- tbl(my_db,"Player_Attributes")
player <- tbl(my_db,"Player")
team_attribute <- tbl(my_db,"Team_Attributes")
team <- tbl(my_db,"Team")
match <- tbl(my_db,"Match")
league <- tbl(my_db,"League")
players_attribute <- data.frame(player_attribute)
players <- data.frame(player)
team_attribute <- data.frame(team_attribute)
team <- data.frame(team)
match <- data.frame(match)
league <- data.frame(league)

## Players

### Get overall ratings of each player

ratings <- players_attribute %>% group_by(player_api_id) %>% filter(overall_rating == max(overall_rating))
top_distinct <- distinct(ratings, .keep_all = TRUE)
distinct_players <- top_distinct[2:5]

### Get the sum of gk attributes according to each player

gk_att <- select(player_attribute, player_api_id, player_fifa_api_id, gk_diving,gk_handling,gk_kicking,gk_positioning,gk_reflexes)
gk_att <- data.frame(gk_att)
gk <- transform(gk_att, gksum=rowSums(gk_att[,3:ncol(gk_att)]))
gk <- gk %>% group_by(player_api_id) %>% filter(gksum == max(gksum))
distinct_goalkeepers <- distinct(gk, .keep_all = TRUE)

### Join for previous tables

add <- full_join(distinct_players,distinct_goalkeepers)

### Get the number of appearances according to each player

home <- select(match,match_api_id,home_team_api_id,home_player_1,home_player_2,home_player_3,home_player_4,home_player_5,home_player_6,home_player_7,home_player_8,home_player_9,home_player_10,home_player_11)
away <- select(match,match_api_id,away_team_api_id,away_player_1,away_player_2,away_player_3,away_player_4,away_player_5,away_player_6,away_player_7,away_player_8,away_player_9,away_player_10,away_player_11)
hp1 <- count(home,home_player_1)
hp2 <- count(home,home_player_2)
hp3 <- count(home,home_player_3)
hp4 <- count(home,home_player_4)
hp5 <- count(home,home_player_5)
hp6 <- count(home,home_player_6)
hp7 <- count(home,home_player_7)
hp8 <- count(home,home_player_8)
hp9 <- count(home,home_player_9)
hp10 <- count(home,home_player_10)
hp11 <- count(home,home_player_11)
colnames(hp1) <- c("player_api_id","count")
colnames(hp2) <- c("player_api_id","count")
colnames(hp3) <- c("player_api_id","count")
colnames(hp4) <- c("player_api_id","count")
colnames(hp5) <- c("player_api_id","count")
colnames(hp6) <- c("player_api_id","count")
colnames(hp7) <- c("player_api_id","count")
colnames(hp8) <- c("player_api_id","count")
colnames(hp9) <- c("player_api_id","count")
colnames(hp10) <- c("player_api_id","count")
colnames(hp11) <- c("player_api_id","count")
j1 <- full_join(hp1,hp2)
j2 <- full_join(j1,hp3)
j3 <- full_join(j2,hp4)
j4 <- full_join(j3,hp5)
j5 <- full_join(j4,hp6)
j6 <- full_join(j5,hp7)
j7 <- full_join(j6,hp8)
j8 <- full_join(j7,hp9)
j9 <- full_join(j8,hp10)
j10 <- full_join(j9,hp11)
ap1 <- count(away,away_player_1)
ap2 <- count(away,away_player_2)
ap3 <- count(away,away_player_3)
ap4 <- count(away,away_player_4)
ap5 <- count(away,away_player_5)
ap6 <- count(away,away_player_6)
ap7 <- count(away,away_player_7)
ap8 <- count(away,away_player_8)
ap9 <- count(away,away_player_9)
ap10 <- count(away,away_player_10)
ap11 <- count(away,away_player_11)
colnames(ap1) <- c("player_api_id","count")
colnames(ap2) <- c("player_api_id","count")
colnames(ap3) <- c("player_api_id","count")
colnames(ap4) <- c("player_api_id","count")
colnames(ap5) <- c("player_api_id","count")
colnames(ap6) <- c("player_api_id","count")
colnames(ap7) <- c("player_api_id","count")
colnames(ap8) <- c("player_api_id","count")
colnames(ap9) <- c("player_api_id","count")
colnames(ap10) <- c("player_api_id","count")
colnames(ap11) <- c("player_api_id","count")
k1 <- full_join(ap1,ap2)
k2 <- full_join(k1,ap3)
k3 <- full_join(k2,ap4)
k4 <- full_join(k3,ap5)
k5 <- full_join(k4,ap6)
k6 <- full_join(k5,ap7)
k7 <- full_join(k6,ap8)
k8 <- full_join(k7,ap9)
k9 <- full_join(k8,ap10)
k10 <- full_join(k9,ap11)
appearance1 <- full_join(j10,k10)
appearance1 <- na.omit(appearance1)
data <- aggregate(count~player_api_id,data=appearance1,FUN=sum)

### Cummulative join for previous tables

add1 <- full_join(add,data)

### Get the number of leagues according to each player

league1 <- select(match,league_id,home_player_1,home_player_2,home_player_3,home_player_4,home_player_5,home_player_6,home_player_7,home_player_8,home_player_9,home_player_10,home_player_11,away_player_1,away_player_2,away_player_3,away_player_4,away_player_5,away_player_6,away_player_7,away_player_8,away_player_9,away_player_10,away_player_11)
league1 <- na.omit(league1)

p1 <- league1 %>% group_by(home_player_1)
p1 <- summarise(distinct(p1,league_id),count = n())

p2 <- league1 %>% group_by(home_player_2)
p2 <- summarise(distinct(p2,league_id),count = n())

p3 <- league1 %>% group_by(home_player_3)
p3 <- summarise(distinct(p3,league_id),count = n())

p4 <- league1 %>% group_by(home_player_4)
p4 <- summarise(distinct(p4,league_id),count = n())

p5 <- league1 %>% group_by(home_player_5)
p5 <- summarise(distinct(p5,league_id),count = n())

p6 <- league1 %>% group_by(home_player_6)
p6 <- summarise(distinct(p6,league_id),count = n())

p7 <- league1 %>% group_by(home_player_7)
p7 <- summarise(distinct(p7,league_id),count = n())

p8 <- league1 %>% group_by(home_player_8)
p8 <- summarise(distinct(p8,league_id),count = n())

p9 <- league1 %>% group_by(home_player_9)
p9 <- summarise(distinct(p9,league_id),count = n())

p10 <- league1 %>% group_by(home_player_10)
p10 <- summarise(distinct(p10,league_id),count = n())

p11 <- league1 %>% group_by(home_player_11)
p11 <- summarise(distinct(p11,league_id),count = n())

p12 <- league1 %>% group_by(away_player_1)
p12 <- summarise(distinct(p12,league_id),count = n())

p13 <- league1 %>% group_by(away_player_2)
p13 <- summarise(distinct(p13,league_id),count = n())

p14 <- league1 %>% group_by(away_player_3)
p14 <- summarise(distinct(p14,league_id),count = n())

p15 <- league1 %>% group_by(away_player_4)
p15 <- summarise(distinct(p15,league_id),count = n())

p16 <- league1 %>% group_by(away_player_5)
p16 <- summarise(distinct(p16,league_id),count = n())

p17 <- league1 %>% group_by(away_player_6)
p17 <- summarise(distinct(p17,league_id),count = n())

p18 <- league1 %>% group_by(away_player_7)
p18 <- summarise(distinct(p18,league_id),count = n())

p19 <- league1 %>% group_by(away_player_8)
p19 <- summarise(distinct(p19,league_id),count = n())

p20 <- league1 %>% group_by(away_player_9)
p20 <- summarise(distinct(p20,league_id),count = n())

p21 <- league1 %>% group_by(away_player_10)
p21 <- summarise(distinct(p21,league_id),count = n())

p22 <- league1 %>% group_by(away_player_11)
p22 <- summarise(distinct(p22,league_id),count = n())

colnames(p1) <- c("player_api_id","league")
colnames(p2) <- c("player_api_id","league")
colnames(p3) <- c("player_api_id","league")
colnames(p4) <- c("player_api_id","league")
colnames(p5) <- c("player_api_id","league")
colnames(p6) <- c("player_api_id","league")
colnames(p7) <- c("player_api_id","league")
colnames(p8) <- c("player_api_id","league")
colnames(p9) <- c("player_api_id","league")
colnames(p10) <- c("player_api_id","league")
colnames(p11) <- c("player_api_id","league")
colnames(p12) <- c("player_api_id","league")
colnames(p13) <- c("player_api_id","league")
colnames(p14) <- c("player_api_id","league")
colnames(p15) <- c("player_api_id","league")
colnames(p16) <- c("player_api_id","league")
colnames(p17) <- c("player_api_id","league")
colnames(p18) <- c("player_api_id","league")
colnames(p19) <- c("player_api_id","league")
colnames(p20) <- c("player_api_id","league")
colnames(p21) <- c("player_api_id","league")
colnames(p22) <- c("player_api_id","league")

j1 <- full_join(p1,p2)
j2 <- full_join(j1,p3)
j3 <- full_join(j2,p4)
j4 <- full_join(j3,p5)
j5 <- full_join(j4,p6)
j6 <- full_join(j5,p7)
j7 <- full_join(j6,p8)
j8 <- full_join(j7,p9)
j9 <- full_join(j8,p10)
j10 <- full_join(j9,p11)
j11 <- full_join(j10,p12)
j12 <- full_join(j11,p13)
j13 <- full_join(j12,p14)
j14 <- full_join(j13,p15)
j15 <- full_join(j14,p16)
j16 <- full_join(j15,p17)
j17 <- full_join(j16,p18)
j18 <- full_join(j17,p19)
j19 <- full_join(j18,p20)
j20 <- full_join(j19,p21)

### Cummulative join for previous tables

add2 <- full_join(add1,j20)
add2

### Removing unncessary columns

add2 <- add2[,-1]

### Removing unncessary columns

add2 <- add2[, -c(4:8)] 

### Removing unncessary columns

add2 <- add2[,-2]

add2

### Renaming columns

colnames(add2) <- c("player_api_id", "overall_rating", "gk_rating", "num_of_appearances", "num_of_leagues")

### Cummulative join for previous tables

final_players <- full_join(players[2:3],add2)

### Writing file in csv

write.csv(final_players, "player.csv")

## Teams

### Get the number of build_up according to each team

build_up_play <- select(team_attribute,team_api_id,team_fifa_api_id,buildUpPlaySpeed,buildUpPlayDribbling,buildUpPlayPassing)
buildsum <- transform(build_up_play, build_sum=rowSums(build_up_play[,3:ncol(build_up_play)]))
buildsum <- buildsum %>% group_by(team_api_id) %>% filter(build_sum == max(build_sum))
distinct_build <- distinct(buildsum, .keep_all = TRUE)

### Get the number of chance_creation according to each team

chance_creation <- select(team_attribute,team_api_id,team_fifa_api_id,chanceCreationPassing,chanceCreationCrossing,chanceCreationShooting)
chancesum <- transform(chance_creation, chance_sum=rowSums(chance_creation[,3:ncol(chance_creation)]))
chancesum <- chancesum %>% group_by(team_api_id) %>% filter(chance_sum == max(chance_sum))
distinct_chance <- distinct(chancesum, .keep_all = TRUE)

### Joining previous two tables

add <- full_join(distinct_build,distinct_chance) 

### Get the number of defence according to each team

defence <- select(team_attribute,team_fifa_api_id,team_api_id,defencePressure,defenceAggression,defenceTeamWidth)
defencesum <- transform(defence, defence_sum=rowSums(defence[,3:ncol(defence)]))
defencesum <- defencesum %>% group_by(team_api_id) %>% filter(defence_sum == max(defence_sum))
distinct_defence <- distinct(defencesum, .keep_all = TRUE)

add <- add[, -c(3:5)]

add <- add[, -c(4:6)]

### Cummulative join of previous tables

add1 <- full_join(add,distinct_defence)

add1 <- add1[, -c(5:7)]

### Get the number of goals according to each team

home_team <- select(match,home_team_api_id,home_team_goal)
away_team <- select(match,away_team_api_id,away_team_goal)
p_home <- home_team %>% group_by(home_team_api_id) %>% summarise(home_goals = sum(home_team_goal))
p_away <- away_team %>% group_by(away_team_api_id) %>% summarise(away_goals = sum(away_team_goal))
colnames(p_home) <- c("team_api_id","home_goals")
colnames(p_away) <- c("team_api_id","away_goals")
total <- full_join(p_home,p_away)
total <- cbind(total[1],rowSums(total[2:3]))
colnames(total) <- c("team_api_id","goals")

### Cummulative join of previous tables

add2 <- full_join(add1,total)

### Removing unncessary columns

add2 <- add2[, -2]

### Cummulative join of previous tables

prefinal <- full_join(team,add2)

### Removing unncessary columns

prefinal <- prefinal[, -1]

### Removing unncessary columns

prefinal <- prefinal[, -2]

### Writing data in csv

write.csv(prefinal, "team.csv")
