==> Exploratory Data Analysis - II/Notebook_3 Activity_2.ipynb <==
## Loading Libraries and Reading CSV files

library(dplyr)
players <- read.csv("player.csv")
teams <- read.csv("team.csv")

## Cleaned Players Table

head(players)

## Cleaned Teams Table

head(teams)

### Query 1 : Which are the top 10 players by overall rating?

overall <- arrange(players, desc(overall_rating))
overall[1:10,]

### Query 2 : Which are the top 10 goalkeepers by sum of gk attributes?

gk <- arrange(players, desc(gk_rating))
gk[1:10,]

### Query 3 : Which are the top 10 players by number of appearances (with any team they played with)?

appeareances <- arrange(players, desc(num_of_appearances))
appeareances[1:10,]

### Query 4 : Which are the top 10 players by number of leagues they played in?

leagues <- arrange(players, desc(num_of_leagues))
leagues[1:10,]

### Query 5 : Which are the top 10 teams by sum of build up play attributes?

build_att <- arrange(teams, desc(build_sum))
build_att[1:10,]

### Query 6 : Which are the top 10 teams by sum of chance creation attributes?

chance_att <- arrange(teams, desc(chance_sum))
chance_att[1:10,]

### Query 7 : Which are the top 10 teams by sum of defense attributes?

defence_att <- arrange(teams, desc(defence_sum))
defence_att[1:10,]

### Query 8 : Which are the top 10 teams by number of scored goals?

goals <- arrange(teams, desc(goals))
goals[1:10,]
